// src/lib/api/generated/models/orgRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * The organization role for a user. Defines their permission level.
 */
export type OrgRole = typeof OrgRole[keyof typeof OrgRole];


export const OrgRole = {
  admin: 'admin',
  member: 'member',
} as const;